requirements.txt
asgiref==3.4.1
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
boto3==1.23.10
botocore==1.26.10
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
distro-info===0.18ubuntu0.18.04.1
Django==3.1.3
django-debug-toolbar==3.2.4
django-filter==21.1
django-model-utils==4.2.0
django-notifications-hq==1.7.0
django-storages==1.12.3
djangorestframework==3.12.2
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
jmespath==0.10.0
jsonfield==3.1.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
mysqlclient==2.0.3
netifaces==0.10.4
PAM==0.4.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
PyGObject==3.26.1
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.4
python-dateutil==2.8.2
python-debian==0.1.32
pytz==2022.1
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
s3transfer==0.5.2
SecretStorage==2.3.1
service-identity==16.0.0
six==1.11.0
sqlparse==0.4.2
ssh-import-id==5.7
swapper==1.3.0
systemd-python==234
todo==0.1
Twisted==17.9.0
typing_extensions==4.1.1
ufw==0.36
urllib3==1.26.9
wrapt==1.14.1
zope.interface==4.3.2
